feat: update ui, band-aid fix for notif bug

+++ src/components/CourseList.tsx
@@ -3,7 +3,7 @@
 import { classEntry } from "@/lib/types";
 import { fetchCourses, tellServer } from "@/server-actions/fetch-courses";
 import { useQuery } from "@tanstack/react-query";
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import CourseInput from "./CourseInput";
 import WatchItem from "./WatchItem";
 import {
@@ -16,29 +16,36 @@ import {
 } from "./ui/card";
 
 const CourseList = () => {
-  const [codes, setCodes] = useState<classEntry[]>([]);
-  const [lastUpdated, setLastUpdated] = useState<Date | null>(null);
+  const notifiedRef = useRef<boolean>(false);
+  const [watched, setWatched] = useState<classEntry[]>([]);
+
+  const { data, dataUpdatedAt, isRefetching } = useQuery({
+    queryKey: ["fetch-codes", watched],
+    queryFn: async () => fetchCourses(watched),
+    refetchInterval: 1 * 60000,
+    refetchIntervalInBackground: true,
+    refetchOnMount: false,
+    refetchOnWindowFocus: false,
+  });
 
   const addClass = (courseCode: string, classCode: string) => {
     const code = Number(classCode);
-    const exists = codes.filter((existing) => existing.code === code);
+    const exists = watched.filter((existing) => existing.code === code);
 
-    if (exists.length > 0) return;
+    if (exists.length > 0 || courseCode.length !== 7) return;
 
     const newCodes = [
-      ...codes,
+      ...watched,
       { code: code, course: courseCode.toUpperCase() },
     ];
 
-    setCodes(newCodes);
+    setWatched(newCodes);
     localStorage.setItem("classCodes", JSON.stringify(newCodes));
   };
 
   const removeClass = (classCode: number) => {
-    const newCodes = codes.filter((code) => code.code !== classCode);
-    console.log(newCodes);
-
-    setCodes(newCodes);
+    const newCodes = watched.filter((code) => code.code !== classCode);
+    setWatched(newCodes);
     localStorage.setItem("classCodes", JSON.stringify(newCodes));
   };
 
@@ -47,54 +54,48 @@ const CourseList = () => {
     const parsed = stored !== null ? JSON.parse(stored) : null;
 
     Notification.requestPermission();
-    setLastUpdated(new Date());
 
     if (parsed) {
-      setCodes(parsed);
+      setWatched(parsed);
     }
   }, []);
 
-  const { data } = useQuery({
-    queryKey: ["fetch-codes", codes],
-    queryFn: async () => {
-      const data = await fetchCourses(codes);
-      setLastUpdated(new Date());
-
-      return data;
-    },
-    refetchInterval: 1 * 60000,
-    refetchIntervalInBackground: true,
-    refetchOnMount: false,
-    refetchOnWindowFocus: false,
-  });
+  useEffect(() => {
+    if (isRefetching) notifiedRef.current = false;
+  }, [isRefetching]);
 
   useEffect(() => {
     if (data !== undefined) {
-      setCodes(data.codes);
+      console.log(data.codes[0]?.status);
+      localStorage.setItem("classCodes", JSON.stringify(data.codes));
 
-      if (data.sendNotif) {
-        tellServer(data.codes);
+      if (!notifiedRef.current && data.sendNotif) {
+        notifiedRef.current = true;
+        tellServer("[USE EFFECT] SendNotif is True, sending Notification!");
+        console.log(data.timestamp?.getTime());
 
         new Notification("Course/s changed status.", {
           body: "One of your watched courses either opened or close, check it out!",
         });
       }
+
+      setWatched(data.codes);
     }
   }, [data]);
 
   return (
-    <div className="flex flex-row w-full min-h-full items-center justify-center gap-4">
+    <div className="flex flex-col w-max min-h-full items-center justify-center gap-4">
       <CourseInput addClass={addClass} />
-      <Card>
+      <Card className="flex flex-col w-full h-1/2">
         <CardHeader>
           <CardTitle>Watchlist</CardTitle>
           <CardDescription>
             {"Classes that you're currently watching are here"}
           </CardDescription>
         </CardHeader>
-        <CardContent>
-          {codes.length !== 0 ? (
-            codes.map((code) => (
+        <CardContent className="overflow-y-auto flex flex-wrap max-w-[64.5rem]">
+          {watched.length !== 0 ? (
+            watched.map((code) => (
               <WatchItem key={code.code} code={code} removeItem={removeClass} />
             ))
           ) : (
@@ -103,7 +104,9 @@ const CourseList = () => {
             </span>
           )}
         </CardContent>
-        <CardFooter className="text-gray-500 text-sm">{`Last Updated: ${lastUpdated?.toLocaleTimeString()}`}</CardFooter>
+        <CardFooter className="text-gray-500 text-sm mt-auto">{`Last Updated: ${new Date(
+          dataUpdatedAt
+        ).toLocaleTimeString()}`}</CardFooter>
       </Card>
     </div>
   );
